+++ pybdsim/Data.py
@@ -13,6 +13,8 @@ from . import _General
 from collections import defaultdict as _defaultdict
 import copy as _copy
 import glob as _glob
+import itertools as _itertools
+import math as _math
 import numpy as _np
 import os as _os
 import pickle as _pickle
@@ -385,10 +387,12 @@ class RebdsimFile(object):
         self.histograms1d = {}
         self.histograms2d = {}
         self.histograms3d = {}
+        self.histograms4d = {}
         self.histogramspy = {}
         self.histograms1dpy = {}
         self.histograms2dpy = {}
         self.histograms3dpy = {}
+        self.histograms4dpy = {}
         self.spectra = _defaultdict(Spectra)
         self._Map("", self._f)
         if convert:
@@ -424,6 +428,7 @@ class RebdsimFile(object):
         h1d = self._ListType(currentDir, "TH1D")
         h2d = self._ListType(currentDir, "TH2D")
         h3d = self._ListType(currentDir, "TH3D")
+        h4d = self._ListType(currentDir, "BDSBH4D")
         for h in h1d:
             name = currentDirName + '/' + h
             name = name.strip('/') # protect against starting /
@@ -442,6 +447,12 @@ class RebdsimFile(object):
             hob = currentDir.Get(h)
             self.histograms[name] = hob
             self.histograms3d[name] = hob
+        for h in h4d:
+            name = currentDirName + '/' + h
+            name = name.strip('/')
+            hob = currentDir.Get(h)
+            self.histograms[name] = hob
+            self.histograms4d[name] = hob
         subDirs = self._ListType(currentDir, "Directory")
         for d in subDirs:
             dName = currentDirName + '/' + d
@@ -479,6 +490,12 @@ class RebdsimFile(object):
             result.append(str(leaves.At(i)))
         return result
 
+    def GetModelTree(self):
+        return self._f.Get('ModelTree')
+
+    def GetModel(self):
+        pass
+
     def ConvertToPybdsimHistograms(self):
         """
         Convert all root histograms into numpy arrays.
@@ -494,6 +511,10 @@ class RebdsimFile(object):
         for path,hist in self.histograms3d.items():
             hpy = TH3(hist)
             self.histograms3dpy[path] = hpy
+            self.histogramspy[path] = hpy
+        for path,hist in self.histograms4d.items():
+            hpy = BDSBH4D(hist)
+            self.histograms4dpy[path] = hpy
             self.histogramspy[path] = hpy
 
     def _PopulateSpectraDictionaries(self):
@@ -793,6 +814,8 @@ class ROOTHist(object):
         self.xlabel = hist.GetXaxis().GetTitle()
         self.ylabel = hist.GetYaxis().GetTitle()
         self.errorsAreErrorOnMean = True
+        self.entries = 0
+        self.errors  = _np.array([]) # avoid problems with error methods
 
     def __getstate__(self):
         """
@@ -957,6 +980,244 @@ class TH3(TH2):
                 for k in range(self.nbinsz):
                     self.contents[i,j,k] = self.hist.GetBinContent(i+1,j+1,k+1)
                     self.errors[i,j,k]   = self.hist.GetBinError(i+1,j+1,k+1)
+
+    def IntegateAlong1Dimension(self, dimension):
+        """
+        Integrate along a dimension returning a new 2D histogram.
+        
+        :param dimension: 'x', 'y' or 'z' dimension to integrate along
+        :type  dimension: str
+
+        returns pybdsim.Data.TH2 instance.
+
+        If the projection is done in z, a 2D histogram of x,y is returned
+        that is the sum of the bins along z. The errors are also calculated.
+
+        For 'x', the 2D histogram is z,y.
+        For 'y', the 2D histogram is z,x.
+        For 'z', the 2D hsitogram is x,y.
+        """
+        if dimension == 'x':
+            h2d = self.hist.Project3D("yze")
+            return TH2(h2d)
+        elif dimension == 'y':
+            h2d = self.hist.Project3D("xze")
+            return TH2(h2d)
+        elif dimension == 'z':
+            h2d = self.hist.Project3D("yxe")
+            return TH2(h2d)
+        else:
+            raise ValueError("dimension can only be one of 'x', 'y', 'z'")
+
+    def IntegateAlong2Dimensions(self, resultDimension):
+        """
+        Integrate along 2 dimensions returning a new 1D histogram along the result dimension
+        
+        :param resultDimension: 'x', 'y' or 'z' dimension to produce 1D histogram along.
+        :type  resultDimension: str
+
+        returns pybdsim.Data.TH1 instance.
+        """
+        if resultDimension == 'x':
+            h1d = self.hist.Project3D('xe')
+            return TH1(h1d)
+        elif resultDimension == 'y':
+            h1d = self.hist.Project3D("ye")
+            return TH1(h1d)
+        elif resultDimension == 'z':
+            h1d = self.hist.Project3D("ze")
+            return TH1(h1d)
+        else:
+            raise ValueError("dimension can only be one of 'x', 'y', 'z'")
+
+    def Slice2DXY(self, index):
+        """
+        Extract a single 2D histogram from an index along the Z dimension.
+        
+        :param index: index in z array of bins to extract, e.g. 0 -> nbinsz-1
+        :type  index: int
+        """
+        if not (0 <= index < self.nbinsz):
+            raise ValueError("index must be in range [0 : "+str(self.nbinsz-1)+"]")
+        self.hist.GetZaxis().SetRange(index+1,index+2)
+        h2d = self.hist.Project3D("yxe")
+        return TH2(h2d)
+
+    def Slice2DXZ(self, index):
+        """
+        Extract a single 2D histogram from an index along the Y dimension.
+        
+        :param index: index in y array of bins to extract, e.g. 0 -> nbinsy-1
+        :type  index: int
+        """
+        if not (0 <= index < self.nbinsz):
+            raise ValueError("index must be in range [0 : "+str(self.nbinsz-1)+"]")
+        self.hist.GetYaxis().SetRange(index+1,index+2)
+        h2d = self.hist.Project3D("xze")
+        return TH2(h2d)
+
+    def Slice2DZY(self, index):
+        """
+        Extract a single 2D histogram from an index along the X dimension.
+        
+        :param index: index in x array of bins to extract, e.g. 0 -> nbinsx-1
+        :type  index: int
+        """
+        if not (0 <= index < self.nbinsz):
+            raise ValueError("index must be in range [0 : "+str(self.nbinsz-1)+"]")
+        self.hist.GetXaxis().SetRange(index+1,index+2)
+        h2d = self.hist.Project3D("yze")
+        return TH2(h2d)
+
+class BDSBH4D():
+    """
+    Wrapper for a BDSBH instance. Converts to numpy data.
+
+    """
+    def __init__(self, hist, extractData=True):
+        # these members are made to be the same as our "ROOTHist" class
+        # even though it isn't inherited (can't be as data different)
+        self.hist   = hist
+        self.name   = hist.GetName()
+        self.title  = hist.GetTitle()
+        self.xlabel = ""
+        self.ylabel = ""
+        self.errorsAreErrorOnMean = True
+        
+        self.nbinsx = hist.GetNbinsX()
+        self.nbinsy = hist.GetNbinsY()
+        self.nbinsz = hist.GetNbinsZ()
+        self.nbinse = hist.GetNbinsE()
+
+        self.xwidths   = _np.zeros(self.nbinsx)
+        self.xcentres  = _np.zeros(self.nbinsx)
+        self.xlowedge  = _np.zeros(self.nbinsx)
+        self.xhighedge = _np.zeros(self.nbinsx)
+        self.xedges    = _np.zeros(self.nbinsx+1)
+
+        self.ywidths   = _np.zeros(self.nbinsy)
+        self.ycentres  = _np.zeros(self.nbinsy)
+        self.ylowedge  = _np.zeros(self.nbinsy)
+        self.yhighedge = _np.zeros(self.nbinsy)
+        self.yedges    = _np.zeros(self.nbinsy+1)
+
+        self.zwidths   = _np.zeros(self.nbinsz)
+        self.zcentres  = _np.zeros(self.nbinsz)
+        self.zlowedge  = _np.zeros(self.nbinsz)
+        self.zhighedge = _np.zeros(self.nbinsz)
+        self.zedges    = _np.zeros(self.nbinsz+1)
+
+        self.ewidths   = _np.zeros(self.nbinse)
+        self.ecentres  = _np.zeros(self.nbinse)
+        self.elowedge  = _np.zeros(self.nbinse)
+        self.ehighedge = _np.zeros(self.nbinse)
+        self.eedges    = _np.zeros(self.nbinse+1)
+
+        self.contents = _np.zeros((self.nbinsx,self.nbinsy,self.nbinsz,self.nbinse))
+        self.errors   = _np.zeros((self.nbinsx,self.nbinsy,self.nbinsz,self.nbinse))
+
+        self._GetBinsInfo(hist)
+
+        if extractData:
+            self._GetContents(hist)
+
+        self.integral = _np.sum(self.contents)
+        # this assumes uncorrelated
+        self.integralError = _np.sqrt((self.errors**2).sum())
+
+    def ErrorsToSTD(self):
+        """
+        Errors are by default the error on the mean. Call this function
+        to multiply by sqrt(N) to convert to the standard deviation.
+        Will automatically only apply itself once even if repeatedly called.
+        """
+        if self.errorsAreErrorOnMean:
+            self.errors *= _np.sqrt(self.entries)
+            self.errorsAreErrorOnMean = False
+        else:
+            pass # don't double apply calculation
+
+    def ErrorsToErrorOnMean(self):
+        """
+        Errors are by default the error on the mean. However, if you used
+        ErrorsToSTD, you can convert back to error on the mean with this
+        function, which divides by sqrt(N).
+        """
+        if self.errorsAreErrorOnMean:
+            pass # don't double apply calculation
+        else:
+            self.errors /= _np.sqrt(self.entries)
+            self.errorsAreErrorOnMean = True
+
+    def _GetBinsInfo(self, hist):
+        x_step = (hist.h_xmax - hist.h_xmin)/hist.h_nxbins
+        for i in range(hist.h_nxbins):
+            self.xwidths[i]   = x_step
+            self.xlowedge[i]  = hist.h_xmin + i*x_step
+            self.xhighedge[i] = self.xlowedge[i] + self.xwidths[i]
+            self.xcentres[i]  = self.xlowedge[i] + self.xwidths[i]/2
+        self.xrange = (self.xlowedge[0],self.xhighedge[-1])
+        self.xedges = _np.append(self.xlowedge, self.xhighedge[-1])
+
+        y_step = (hist.h_ymax - hist.h_ymin) / hist.h_nybins
+        for i in range(hist.h_nybins):
+            self.ywidths[i]   = y_step
+            self.ylowedge[i]  = hist.h_ymin + i*y_step
+            self.yhighedge[i] = self.ylowedge[i] + self.ywidths[i]
+            self.ycentres[i]  = self.ylowedge[i] + self.ywidths[i] / 2
+        self.yrange = (self.ylowedge[0],self.yhighedge[-1])
+        self.yedges = _np.append(self.ylowedge, self.yhighedge[-1])
+            
+        z_step = (hist.h_zmax - hist.h_zmin) / hist.h_nzbins
+        for i in range(hist.h_nzbins):
+            self.zwidths[i]   = z_step
+            self.zlowedge[i]  = hist.h_zmin + i*z_step
+            self.zhighedge[i] = self.zlowedge[i] + self.zwidths[i]
+            self.zcentres[i]  = self.zlowedge[i] + self.zwidths[i] / 2
+        self.zrange = (self.zlowedge[0],self.zhighedge[-1])
+        self.zedges = _np.append(self.zlowedge, self.zhighedge[-1])
+
+        if hist.h_escale == 'log':
+            e_step = (_math.log10(hist.h_emax) - _math.log10(hist.h_emin)) / hist.h_nebins
+            for i in range(hist.h_nebins):
+                self.elowedge[i]  = hist.h_emin * 10 ** (i * e_step)
+                self.ehighedge[i] = hist.h_emin * 10 ** ((i+1) * e_step)
+                self.ewidths[i]   = self.ehighedge[i] - self.elowedge[i]
+                self.ecentres[i]  = self.elowedge[i] + self.ewidths[i] / 2
+
+        if hist.h_escale == 'linear':
+            e_step = (hist.h_emax - hist.h_emin) / hist.h_nebins
+            for i in range(hist.h_nebins):
+                self.ewidths[i]  = e_step
+                self.elowedge[i]  = hist.h_emin + i * e_step
+                self.ehighedge[i] = self.elowedge[i] + self.ewidths[i]
+                self.ecentres[i]  = self.elowedge[i] + self.ewidths[i] / 2
+
+        if hist.h_escale == 'user':
+            for i in range(hist.h_nebins):
+                self.elowedge[i]  = hist.h_ebinsedges.at(i)
+                self.ehighedge[i] = hist.h_ebinsedges.at(i+1)
+                self.ewidths[i]   = hist.h_ebinsedges.at(i+1)-hist.h_ebinsedges.at(i)
+                self.ecentres[i]  = self.elowedge[i] + 0.5*self.ewidths[i]
+
+        self.erange = (self.elowedge[0],self.ehighedge[-1])
+        self.eedges = _np.append(self.elowedge, self.ehighedge[-1])
+
+
+    def _ToNumpy(self, hist, hist_type="h"):
+        histo4d = _np.zeros((hist.h_nxbins, hist.h_nybins, hist.h_nzbins, hist.h_nebins))
+        ho = getattr(hist, hist_type)
+        for x, y, z, e in _itertools.product(range(hist.h_nxbins),
+                                             range(hist.h_nybins),
+                                             range(hist.h_nzbins),
+                                             range(hist.h_nebins)):
+            histo4d[x, y, z, e] = ho.at(x, y, z, e)
+
+        return histo4d
+
+    def _GetContents(self,hist):
+        self.contents = self._ToNumpy(hist)
+        self.errors   = self._ToNumpy(hist, hist_type="h_err")
 
 
 class _SamplerData(object):
@@ -1373,7 +1634,7 @@ class EventSummaryData(EventInfoData):
         self._getData(interface, info, eventTree)
 
 def GetApertureExtent(apertureType, aper1=0, aper2=0, aper3=0, aper4=0):
-    apertureType = apertureType.lower()
+    apertureType = str(apertureType).lower()
 
     if apertureType == "":
         return 0,0
@@ -1406,14 +1667,14 @@ class ApertureInfo(object):
     Simple class to hold aperture parameters and extents.
     """
     def __init__(self, apertureType, aper1, aper2=0, aper3=0, aper4=0, offsetX=0, offsetY=0):
-        self.apertureType = apertureType
+        self.apertureType = str(apertureType) # maybe not a python str type from data
         self.aper1    = aper1
         self.aper2    = aper2
         self.aper3    = aper3
         self.aper4    = aper4
         self.offsetX  = offsetX
         self.offsetY  = offsetY
-        self.x,self.y = GetApertureExtent(apertureType, aper1, aper2, aper3, aper4)
+        self.x,self.y = GetApertureExtent(self.apertureType, aper1, aper2, aper3, aper4)
 
 class ModelData(object):
     def __init__(self, data):